Support multiple concurrent kubernetes versions
This change extends the playbooks to allow multiple kubernetes versions to
exist in a cluster concurrently, to support gradual rolling upgrades. To
achieve this the simple variables:

	kubernetes_version: 1.4.8-00
	kubernetes_cni_version: 0.3.0.1-07a8a2-00
	kubernetes_image_version: v1.4.8

Are replaced by a map and a default selector:

	defaultk8s: 1.4.8
	kubernetes_versions:
      1.4.8:
        kubelet : 1.4.8-00
        cni: 0.3.0.1-07a8a2-00
        images: v1.4.8

Furthermore, this default can be overriden on a per-host basis in the inventory
file by specifying a `hostk8s=x.y.z` variable that selects a different key from
the map. With this mechanism in place, it is possible to upgrade a cluster
incrementally, applying the site playbook to all hosts at each stage, ensuring
the site configuration specification does not diverge from the cluster state.